speedspacechart: fix blinking in speed limit tags by importing images only once
